==> docs/6_CMDynamicsSimulations.Rmd <==
---
title: "COMPMUT Dynamics 6: Simulations and analysis part 2"
date: "compiled Aug 2024"
author: "jpjh"
output: rmarkdown::github_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
rm(list=ls())
library(tidyverse)
library(ggplot2)
library(deSolve)
library(rootSolve)
library(knitr)
source("./functions/theme_pub.r")
theme_github <- function(x){
  theme_pub(base_size=12)
}
theme_set(theme_github())
```

## Numerical simulations of compensatory mutation dynamics

As described in [`4_PlaCMSimulations.Rmd`](./docs/4_PlaCMSimulations.md), a general ODE model was developed simulating competition between plasmid-free, wild-type plasmid containing, and two different CMs. 

The model is specified in `./model/COMPMOD_model.R`, and was used to generate plots describing longitudinal dynamics of the system, to complement the parameter space phase diagrams generated by analysis (Figure 1). 

Recall:

- Z_f = no plasmid (i.e. SBW25)
- X_0 = plasmid-carrier (i.e. SBW25(pQBR57))
- X_1 = plasmid-carrier with CM type 1 (i.e. SBW25(pQBR57::plaCM))
- X_2 = plasmid-carrier with CM type 2 (i.e. SBW25::chrCM(pQBR57))
- X_0t = transconjugant from X_0
- X_1t = transconjugant from X_1 (i.e. wild-type recipients of pQBR57::plaCM)
- X_2t = transconjugant from X_2 (i.e. wild-type recipients of unameliorated plasmid from CM type 2)

### Identify suitable model parameters

As before, the benchmark for the system is the parameters measured in `3_ParameterEstimation.Rmd`.

```{r}
(parameters <- read.table("./data/3_Parameters.txt", header=TRUE)) %>% kable(digits = 20)
```

As described in [`4_PlaCMSimulations.Rmd`](./docs/4_PlaCMSimulations.md), this puts us in the regime where both the wild-type plasmid and plasmids from CM strains can invade and dominate. 

Note: defining 'invasion' if gamma > μ(α-β)/(α-μ). 'Domination' if gamma > μ(α-β)/(β-μ).

To explore key features of the phase diagrams, we are also interested in cases where the wild-type and compensated plasmids can be lost, can invade, and can dominate. 

```{r}
param_values <- parameters %>% select(-sd) %>% 
  pivot_wider(names_from = "parameter", values_from = "value") %>%
  mutate(gamma_ = gamma * K,
         beta_P = alpha * beta,
         beta_C = alpha * beta_chrCM,
         beta_Q = alpha * beta_plaCM)

(comparison_table <- data.frame(
  comparison = c("no","chrCM","plaCM"),
  beta = c(param_values$beta_P,param_values$beta_C,param_values$beta_Q),
  gamma = c(rep(param_values$gamma_,3)),
  alpha = param_values$alpha,
  mu = param_values$mu) %>%
  mutate(inv_threshold = (mu * (alpha - beta)/(alpha - mu)),
         inv_ratio = gamma / inv_threshold,
         invasion = ifelse(gamma > inv_threshold, "YES", "no"),
         dom_threshold = (mu * (alpha - beta)/(beta - mu)),
         dom_ratio = gamma / dom_threshold,
         domination = ifelse(gamma > dom_threshold, "YES", "no"))) %>% 
  select(-alpha, -mu) %>% kable()
```

To achieve this, we focus on the plaCM parameters for consistency across the panels. We introduce alternative values of gamma that affect the fate of the compensated and non-compensated plasmids.

Note that other parameters can be explored using the interactive Shiny app at https://jpjh.shinyapps.io/COMPMOD_shiny/. 

#### Dynamics for chrCM

Models run for chrCM begin with cmod.

In this model, we set up the parameters so CM 1 is not present, and fitness and conjugation rate for the CM 2 transconjugants are the same as the wild-type plasmid-carrier. Note that we will use the parameters (i.e. costs) from plaCM, for consistency across plots.

`parameters_cmod0` contains the default parameters for the run. The gamma_X_0/0t and gamma_X_1/1t are kept open for modification.

```{r}
parameters_mod0 <- c(alpha_Z_f = param_values$alpha,
                     alpha_X_0 = param_values$beta_P,
                     alpha_X_1 = param_values$beta_Q,
                     alpha_X_2 = param_values$beta_Q,
                     alpha_X_0t = param_values$beta_P,
                     alpha_X_1t = param_values$beta_Q,
                     alpha_X_2t = param_values$beta_P,
                     eta_Z_f = 0,
                     eta_X_0 = 0,
                     eta_X_1 = 0,
                     eta_X_2 = 0,
                     eta_X_0t = 0,
                     eta_X_1t = 0,
                     eta_X_2t = 0,
                     K = param_values$K,
                     mu = param_values$mu)

parameters_cmod0 <- c(parameters_mod0,
                      gamma_X_1 =  param_values$gamma,
                      gamma_X_1t = param_values$gamma)
```

Generate the parameters for use. 

- Green = both plasmid-bearing types can dominate
- Red = CM cannot dominate, wt cannot invade
- Blue = CM cannot dominate, wt can invade
- Orange = CM cannot dominate, wt can dominate

Set up these parameters as variables for clarity.

Assume that the CM at least halves conjugation rate, to ensure we are always in the non-trivial region of parameter space. 

```{r}
wt_cannot_dominate <- 1/3.1
wt_cannot_invade <- 1/3.5
cm_cannot_dominate <- 1/10.5
cm_cannot_invade <- 1/12


parameters_cmod0_green <- c(parameters_cmod0,
                            gamma_X_0 = param_values$gamma,
                            gamma_X_0t = param_values$gamma,
                            gamma_X_2 = param_values$gamma * 0.5,
                            gamma_X_2t = param_values$gamma)

parameters_cmod0_red   <- c(parameters_cmod0,
                            gamma_X_0 = param_values$gamma * wt_cannot_invade,
                            gamma_X_0t = param_values$gamma * wt_cannot_invade,
                            gamma_X_2 = param_values$gamma * cm_cannot_invade,
                            gamma_X_2t = param_values$gamma * wt_cannot_invade)

parameters_cmod0_blue   <- c(parameters_cmod0,
                            gamma_X_0 = param_values$gamma * wt_cannot_dominate,
                            gamma_X_0t = param_values$gamma * wt_cannot_dominate,
                            gamma_X_2 = param_values$gamma * cm_cannot_invade,
                            gamma_X_2t = param_values$gamma * wt_cannot_dominate)

parameters_cmod0_orange   <- c(parameters_cmod0,
                            gamma_X_0 = param_values$gamma,
                            gamma_X_0t = param_values$gamma,
                            gamma_X_2 = param_values$gamma * cm_cannot_invade,
                            gamma_X_2t = param_values$gamma)

parameters_cmod0_pink   <- c(parameters_cmod0,
                             gamma_X_0 = param_values$gamma * wt_cannot_dominate,
                             gamma_X_0t = param_values$gamma * wt_cannot_dominate,
                             gamma_X_2 = param_values$gamma,
                             gamma_X_2t = param_values$gamma * wt_cannot_dominate)
```


Run as a continuous model, i.e. mu = 0.04125 and dilFac = 1 (i.e. no dilution at each 'transfer', washout is continuous). Run for a long time to get approximate steady-state dynamics.

```{r}
source("./model/COMPMOD_model.R")

numTransfers <- 1536
transferTime <- 24
dilFac <- 1

times <- seq(0, transferTime*numTransfers, by = 0.1) 

startfrac <- unname(parameters_cmod0["K"] * 0.01)   # the population size at the start, 
            # as estimated as ~ starting overnight cultures at saturation

initial_state <- c(Z_f = startfrac,   
                 X_0 = startfrac,  
                 X_1 = 0,  
                 X_2 = startfrac,
                 X_0t = 0,
                 X_1t = 0,
                 X_2t = 0)

state_mod01 <- initial_state * c(0.5,0.25,0,0.25,0,0,0)

event <- data.frame(var = rep(c("Z_f", "X_0", "X_1", "X_2", "X_0t", "X_1t", "X_2t"), numTransfers),
                    time = rep(seq(transferTime,transferTime*numTransfers,by=transferTime),each=7), 
                    value = c(dilFac), # set to 1 for the continuous model
                    method = c("mult"))

runModel <- function(state, func, times, parms, events, cm="", param_region="") {
  data.frame(ode(y = state, times = times,
                 func = func, parms = parms,
                 events = list(data = events))) %>%
    pivot_longer(cols=-time, names_to = "subpop", values_to = "density") %>%
    mutate(subpop = factor(subpop, levels=c('X_0','X_0t','X_1','X_1t','X_2','X_2t','Z_f'))) %>%
    group_by(time, subpop) %>% summarise(tot = sum(density), .groups='drop_last') %>%
    mutate(fraction = tot/sum(tot),
           cm = cm,
           param_region = param_region)
}

cmod0_green <- runModel(state = state_mod01,
                        func = COMPMOD_model,
                        times = times,
                        events = event,
                        parms = parameters_cmod0_green,
                        cm = "chrCM",
                        param_region = "green")

cmod0_red <- runModel(state = state_mod01,
                        func = COMPMOD_model,
                        times = times,
                        events = event,
                        parms = parameters_cmod0_red,
                        cm = "chrCM",
                        param_region = "red")

cmod0_blue <- runModel(state = state_mod01,
                        func = COMPMOD_model,
                        times = times,
                        events = event,
                        parms = parameters_cmod0_blue,
                        cm = "chrCM",
                        param_region = "blue")

cmod0_orange <- runModel(state = state_mod01,
                        func = COMPMOD_model,
                        times = times,
                        events = event,
                        parms = parameters_cmod0_orange,
                        cm = "chrCM",
                        param_region = "orange")

cmod0_pink_high <- runModel(state = state_mod01,
                            func = COMPMOD_model,
                            times = times,
                            events = event,
                            parms = parameters_cmod0_pink,
                            cm = "chrCM",
                            param_region = "pink")

cmod0_pink_low <- runModel(state = initial_state * c(0.5,0.49995,0,0.00005,0,0,0),
                            func = COMPMOD_model,
                            times = times,
                            events = event,
                            parms = parameters_cmod0_pink,
                            cm = "chrCM",
                            param_region = "pink (low frequency)")
```

Plot each model together.

```{r}
(left_panel <- bind_rows(cmod0_green, cmod0_red, cmod0_blue, cmod0_orange, cmod0_pink_high) %>% 
  filter(time %in% seq(0,transferTime*numTransfers,transferTime)) %>%
  ggplot(aes(time/10000, fraction, fill = subpop)) +
  geom_area() +
  scale_fill_manual(values=c("#ccb333","#e6d898","#44AA99","#92d3c8",
                             "#882255","#CC6677","#DDDDDD"),
                    breaks=c("X_0","X_0t","X_1","X_1t","X_2","X_2t","Z_f"),
                    labels=c("wild-type plasmid donor","wild-type plasmid transconjugant",
                             "plaCM donor","plaCM transconjugant",
                             "chrCM donor","chrCM transconjugant",
                             "plasmid-free"),
                    name="subpopulation", drop=TRUE) +
  scale_x_continuous(name = "time (/10000)", breaks=c(0,1,2,3)) + 
  facet_grid(cm~param_region) +
  theme(legend.position="bottom"))
```

Look also at the plot for pink with low starting frequency.

```{r}
(pink_plots <- bind_rows(cmod0_pink_low, cmod0_pink_high) %>%
  filter(time %in% seq(0,transferTime*numTransfers,transferTime)) %>%
  ggplot(aes(time/10000, fraction, fill = subpop)) +
  geom_area() +
  scale_fill_manual(values=c("#ccb333","#e6d898","#44AA99","#92d3c8",
                             "#882255","#CC6677","#DDDDDD"),
                    breaks=c("X_0","X_0t","X_1","X_1t","X_2","X_2t","Z_f"),
                    labels=c("wild-type plasmid donor","wild-type plasmid transconjugant",
                             "plaCM donor","plaCM transconjugant",
                             "chrCM donor","chrCM transconjugant",
                             "plasmid-free"),
                    name="subpopulation", drop=TRUE) +
  scale_x_continuous(name = "time (/10000)", breaks=c(0,1,2,3)) + 
  facet_grid(cm~param_region) +
  theme(legend.position="bottom", strip.text.y=element_blank()))
```


#### Dynamics for plaCM

Models run for plaCM begin with pmod.

In this model, we set up the parameters so CM 1 is not present, and fitness and conjugation rate for the CM 2 transconjugants are the same as the wild-type plasmid-carrier. Note that we will use the parameters (i.e. costs) from plaCM, for consistency across plots.

`parameters_cmod0` contains the default parameters for the run. The gamma_X_0/0t and gamma_X_1/1t are kept open for modification.

```{r}
parameters_pmod0 <- c(parameters_mod0,
                      gamma_X_2 =  param_values$gamma,
                      gamma_X_2t = param_values$gamma)
```

Generate the parameters for use. 

- Green = both plasmids can dominate
- Red = CM cannot invade, wt cannot invade
- Blue = CM cannot invade, wt can invade
- Orange = CM cannot invade, wt can dominate

Set up these parameters as variables for clarity.

Assume that the CM at least halves conjugation rate, as above. 

Here, as we are looking at plaCM, the transconjugants inherit the same modifier.

```{r}
parameters_pmod0_green <- c(parameters_pmod0,
                            gamma_X_0 = param_values$gamma,
                            gamma_X_0t = param_values$gamma,
                            gamma_X_1 = param_values$gamma * 0.5,
                            gamma_X_1t = param_values$gamma * 0.5)

parameters_pmod0_red   <- c(parameters_pmod0,
                            gamma_X_0 = param_values$gamma * wt_cannot_invade,
                            gamma_X_0t = param_values$gamma * wt_cannot_invade,
                            gamma_X_1 = param_values$gamma * cm_cannot_invade,
                            gamma_X_1t = param_values$gamma * cm_cannot_invade)

parameters_pmod0_blue   <- c(parameters_pmod0,
                            gamma_X_0 = param_values$gamma * wt_cannot_dominate,
                            gamma_X_0t = param_values$gamma * wt_cannot_dominate,
                            gamma_X_1 = param_values$gamma * cm_cannot_invade,
                            gamma_X_1t = param_values$gamma * cm_cannot_invade)

parameters_pmod0_orange   <- c(parameters_pmod0,
                            gamma_X_0 = param_values$gamma,
                            gamma_X_0t = param_values$gamma,
                            gamma_X_1 = param_values$gamma * cm_cannot_invade,
                            gamma_X_1t = param_values$gamma * cm_cannot_invade)

parameters_pmod0_purple   <- c(parameters_pmod0,
                            gamma_X_0 = param_values$gamma * wt_cannot_invade,
                            gamma_X_0t = param_values$gamma * wt_cannot_invade,
                            gamma_X_1 = param_values$gamma * cm_cannot_dominate,
                            gamma_X_1t = param_values$gamma * cm_cannot_dominate)
```


Run as a continuous model, i.e. mu = 0.04125 and dilFac = 1 (i.e. no dilution at each 'transfer', washout is continuous). 

```{r}
initial_state <- c(Z_f = startfrac,   
                 X_0 = startfrac,  
                 X_1 = startfrac,  
                 X_2 = 0,
                 X_0t = 0,
                 X_1t = 0,
                 X_2t = 0)

state_mod01 <- initial_state * c(0.5,0.25,0.25,0,0,0,0)

pmod0_green <- runModel(state = state_mod01,
                        func = COMPMOD_model,
                        times = times,
                        events = event,
                        parms = parameters_pmod0_green,
                        cm = "plaCM",
                        param_region = "green")

pmod0_red <- runModel(state = state_mod01,
                        func = COMPMOD_model,
                        times = times,
                        events = event,
                        parms = parameters_pmod0_red,
                        cm = "plaCM",
                        param_region = "red")

pmod0_blue <- runModel(state = state_mod01,
                        func = COMPMOD_model,
                        times = times,
                        events = event,
                        parms = parameters_pmod0_blue,
                        cm = "plaCM",
                        param_region = "blue")

pmod0_orange <- runModel(state = state_mod01,
                        func = COMPMOD_model,
                        times = times,
                        events = event,
                        parms = parameters_pmod0_orange,
                        cm = "plaCM",
                        param_region = "orange")

pmod0_purple <- runModel(state = state_mod01,
                        func = COMPMOD_model,
                        times = times,
                        events = event,
                        parms = parameters_pmod0_purple,
                        cm = "plaCM",
                        param_region = "purple")
```

Plot each model together.

```{r}
(right_panel <- bind_rows(pmod0_green, pmod0_red, pmod0_blue, pmod0_orange, pmod0_purple) %>% 
  filter(time %in% seq(0,transferTime*numTransfers,transferTime)) %>%
  ggplot(aes(time/10000, fraction, fill = subpop)) +
  geom_area() +
  scale_fill_manual(values=c("#ccb333","#e6d898","#44AA99","#92d3c8",
                             "#882255","#CC6677","#DDDDDD"),
                    breaks=c("X_0","X_0t","X_1","X_1t","X_2","X_2t","Z_f"),
                    labels=c("wild-type plasmid donor","wild-type plasmid transconjugant",
                             "plaCM donor","plaCM transconjugant",
                             "chrCM donor","chrCM transconjugant",
                             "plasmid-free"),
                    name="subpopulation", drop=TRUE) +
  scale_x_continuous(name = "time (/10000)", breaks=c(0,1,2,3)) + 
  facet_grid(cm~param_region) +
  theme(legend.position="bottom"))
```

Plot together as a single figure.

```{r}
library(patchwork)

short_plot <- left_panel + right_panel + 
  plot_layout(guides = "collect") & theme(legend.position = 'bottom', strip.text.y=element_blank())

tiff("./figs/Fig9.tiff", width=10.8, height=3.6, units="in", res=300)
short_plot
dev.off()

tiff("./figs/Fig10.tiff", width=2.8, height=3.6, units="in", res=300)
pink_plots
dev.off()
```

Final assembly outside of R.

---

**[Back to index.](../README.md)**
